requirements.txt
fastapi
uvicorn
websockets
scikit-learn
pandas
groq
python-dotenv
sqlalchemy
aiosqlite
pydantic